[MAJOR] replace the wait-queue linked list with an rbtree.
This is a backport of the patch from Sin Yu which was merged in 1.3.
It makes use of an rbtree for the wait queue, eliminating the high
CPU usage when using heterogenous timeouts in the configuration with
thousands of simultaneous connections. This code works fairly well
in 1.3, so it was worth back-porting it to 1.2 for all users of the
stable version.