[MEDIUM] add support for -L ip:port on the command line
This adds one listener with one server after the last entries.
It may be used to quickly set up a listening port from an empty
configuration. The maxconn is taken from global.maxconn, and the
clitimeout is arbitrarily set to 10 seconds.

This also fixes a bug causing busy loops upon first connection
when a listener's maxconn was not set.