MEDIUM: enable zero-copy sending using vmsplice()+tee()+splice()
The principle is to first create a large buffer using vmsplice(), then
use tee() to refill a second buffer where splice() gets the data to be
sent.

Right now, output buffers are not realigned but it is not that big of
an issue.

Two new global settings have been added, "nosplice" (to disable the
feature) and "pipesize xxx" to set the pipe buffer size to xxx bytes.