MAJOR: really implement keep-alive
The keep-alive support is ugly and will not work well for POSTs nor
pipelining but at least it's able to wait for a new request after
having sent the last bytes of the response. HTTP/1.1 presets KA,
connection: c* clears it and connection: k* sets it.

This also required to disable TCP_CORK which was already obsoleted
by MSG_MORE.