MINOR: support hiding the body length with ?b=
For testing it's sometimes useful not to advertise the body length
with close responses. Passing "b=0" will cause the Content-Length
header to be renamed to "X-Content-Length" so that it's possible to
debug using it but it's not parsed by the client. Of course it's not
a good idea to do this with keep-alive.