MINOR: support extra parameter delimiters like comma and slash
The "&" character is annoying on the command line as it requires
escaping, just like the semi-colon. This patch adds support for the
semi-colon, the question mark, the slash and the comma, the latter two
being more command-line friendly. So it's now possible to pass
/?b=0/s=10k without problem. Supporting the question mark eases the
creation of query strings in scripts by concatenating series of
"?arg=value".