MEDIUM: consume the whole request body before starting to process the request
We used not to properly deal with POST requests by responding immediately
then draining remaining data for some time. It's causing lots of
complications for a number of tests cases. This patch implements consumption
of the request body before request processing. It eats all headers and the
advertised content-length and then falls back to waiting (if needed) and
producing response data. This means that advertising a long POST will really
result in this POST to be consumed now.

A new CL_STBODY state was inserted between CL_STHEADERS and CL_STWAIT for
this.