MEDIUM: make the output headers more modular
The way output headers are emitted became a real mess with strings
containing empty headers or with "X-" being prepended in front of
certain names. In addition, the content-length may be emitted even
when keep-alive is enabled, which just freezes connections.

Let's refactor the header emission code. It turns out that simplifying
it allows to remove many snprintf() calls and to replace them with
memcpy() of static strings. As a side effect, the request rate gained
about 10%.