MINOR: support the 'r' "multiple" to sizes to add randomness
When a size is suffixed with 'r', a random value between the 0 and the
value-1 is used. The multiplicator suffixes (k,m,g) are applied after
the random so that s=10kr will have 10 discrete values 1k apart while
s=1024r will have 1024 different values. It may even be used for
booleans.