MINOR: support /?B to limit the amount of incoming body before responding
Sometimes in order to reproduce certain bugs it's useful to be able to
respond before reading all the body. This patch adds "/?B=" to set the
amount of body that is expected to be read before responding. The rest
will be mostly dropped and may result in a TCP RST if truncated too early
but this actually matches what some servers do on redirects. Note that 0
is a valid value (i.e. respond immediatley).