MINOR: make the "pieces" mode wait less
The random delay was between 1 and 256 ms between two pieces, with an
average around 16ms. That's way too slow and basically unusable. Not
waiting doesn't work either because packets still have the time to be
re-aggregated in the network stack between two consecutive send() call,
despite TCP_NODELAY. Testing shows that using a random value between
0 and 15ms seems to give good results, with 1k conn showing 1.3 Gbps
and 460kpps over the loopback, causing the client to perform reads of
varying sizes.