MINOR: make it clear in the help message that integers are for use everywhere
httpterm already suppors units and random in every value including booleans
so that it's possible to write "c=30r" to have roughly 3% non-cacheable,
since only the zero integer value is false and any value above is true. This
is also very convenient for keep-alive. Let's make this clearer in the help
message.