MEDIUM: rely on req_maxbody instead of req_body for early responses
The ability to respond before the end of data is nice, but it was done
a bit too much as a hack as it would limit the apparent req_body value
and cause trouble to keep-alive on multiple posts in this case.

Instead now we're updating req_maxbody to reflect the proper share of
req_body that must be read before responding and only rely on this one.
Once done, req_body still serves to know whether we've read everything
or not and event_cli_read() to only drain what's necessary.