BUG/MEDIUM: fix request pipelining
Keep-alive support was very rudimentary when it was added, but once in
a while I get trapped breaking h2load -m, and the fix is not that hard
so let's do it now.

We must not consume the rest of the input buffer's data when switching
to the data mode nor can we do it in keep-alive, we have to keep it
intact and possibly to realign the contents. This also implies that only
the exact amount of advertised body is consumed and never more.
After a response is finished, we realign the request buffer so that it
remains unwrapped as expected by the parser. We could in theory avoid
the memmove as long as the reserve is not touched but until we parse
the request we don't know if the reserve will be touched by extra data,
and the pipelining is rare enough not to be an interesting optimization
target anyway.