MINOR: epoll: use a low value instead of 64 for the maximum epoll events
epoll_wait() was still accepting up to maxfd events due to the history
around the select() call that slipped there. Tests show that using a
very low value work best. Values between 8 and 16 are optimal for request
rate and values 64-128 are optimal for bit rate. 64 gives a good balance
at around 98-99% of the max of each, so let's use that.