MEDIUM: make the maxconn default to available FDs
Maxconn used to default to -n or 2000, which easily fools the user. Now
we'll set it to the number of available FDs after listeners have been
subtracted (hard limit). This will allow the largest possible value by
default as long as -n is not forced.

The DEFAULT_MAXCONN is now only used when ulimit -n is not readable or
too low. Other instances will never match anyway since they're used as
fallbacks from cfg_maxconn (e.g. for proxy maxconn).