BUG/MEDIUM: fix chunked response ending.
The chunked encoding code was wrong about the message termination, it
would not send the final CRLF and was assuming that 0CRLF is sufficient.
Modifying it was a bit tricky because all the code is built on the
assumption that chunk sizes are multiples of 3 bytes, so a few special
cases had to be fitted there.

In order not to have to deal with splice-based encoding of the last
5-bytes message, we simply refrain from using splice for messages smaller
than MIN_SPLICE (4kB since below that this will induce memcpy anyway).

Now both splice and send appear to work well. Quite often a 5-byte packet
is seen for the final chunk. It seems that SPLICE_F_MORE and send() do
not always cope optimally. Maybe using a dedicated pipe for all the final
chunks and concatenating them would be better.