MINOR: work around recent glibc bugs which break FD_SET*
Some glibc version started to purposely break select(), FD_SET, FD_CLR,
FD_ISSET and to proudly document the breakage as intentional (!). While
older versions and virtually all operating systems are fine with either
an fd_set array or redefining FD_SETSIZE, new glibcs do not support this
and break a bunch 20-years old programs by deciding to crash during the
FD_* operations.

This patch replaces them with open-coded ones that do the same using
integers. This was not tested on big-endian systems but given that
haproxy used to do that a long time ago on such systems, it should not
be an issue.