MINOR: add an optional argument to limit the send buf size
When dealing with tens of thousands of concurrent connections, it can
become difficult for the system to find memory for socket buffers,
especially when running all the processes locally on a laptop, where
all rxbuf and txbuf add up to the amount of RAM consumed by TCP. In
this case, the only solution is to limit the send buf size. While it's
possible with splice(), splice() doesn't always work at such loads due
to the lack of pipes.

Let's just add an argument to set sndbuf on the output socket. This
argument is -B and it takes a size in bytes. For the loopback, small
sizes are sufficient (e.g. 16kB already makes it trivial to send
20+ Gbps over the loopback using only 2 cores).