BUG/MEDIUM: fix SPLICE_F_MORE in keep-alive
When keep-alive was implemented by commit ded7a57 ("MAJOR: really
implement keep-alive"), splice() was adjusted to modulate SPLICE_F_MORE
depending on keep-alive and end of object being reached or not. The
problem is that it used to continue to push as much as possible there
(s->to_write) instead of limiting itself to max, and the condition
caused SPLICE_F_MORE to be pushed on the last block, resulting in low
performance for object sizes close to the SPLICE buffer size (typically
66-68kB had horrible performance).

Between the two approaches (limiting splice() to max with SPLICE_F_MORE
or limiting it to the larger to_write without F_MORE, tests show that
the former definitely is faster by about 15%, and it's also cleaner
anyway as it works like the send() path that respects max.