MINOR: add basic support for Etag production
When testing caches, Etag is really missing. Here when enabled (e>0),
the request line is checksummed (without the method) and this is used
to emit a value in the etag header. This is very rudimentary but we
don't need more actually.

It is not enabled by default because summing the words and emitting the
output using snprintf() cost quite a bit of CPU.