BUG/MEDIUM: set the proxy's maxconn separately
The global.maxconn that is calculated by default based on RLIMIT_NOFILE
was attempted to be applied as the defaul proxy's maxconn, except that
this is done before global.maxconn is even known, resulting in the default
value falling back to DEFAULT_MAXCONN. Because of this, the extraneous FDs
are permitted on the process (e.g. over multiple listeners) but not in a
single proxy, causing accept delays.

Since this value is not used beyond setting an accept limit (and the
listen backlog), let's just use a high default value of 1<<30 as the
limit. It will no longer harm the ability to accept connections.

Thanks to Amaury Denoyelle for reporting this problem.