MINOR: make /?b also support forcing the advertised content-length
This is useful to trigger truncated responses: advertise a larger
content-length than will be sent. For this we change /?b to use the
default C-L if <0, and to force a value if >0. There's no way to force
a zero value.

Combined with keep-alive, a longer content-length than expected will
cause a timeout and present a truncated object to the client. A shorter
content-length will return more data than advertised and will place
garbage on the expectedly idle connection.

Combined with close (/?C=1) the longer content-length will also trigger
an immediate close showing an immediately truncated object.