MINOR: support a pause after the end of the transfer
The new 'P' argument allows to pause a certain number of milliseconds
after the end of the response. With keep-alive, it will only pause before
switching to the next request, causing artificial delays in processing
the next one. But with close mode (C=1) and content-length disabled (b=0),
it will cause a delay before closing at the end of the data, which can be
useful to prevent the FIN from being merged with the last messages.