OPTIM: get rid of the last few snprintf()
snprintf() is super slow, it's taking 35% of the CPU alone. It's much
more efficient to build the response manually using memcpy() and ulltoa(),
despite the increased complexity. With this change, the request rate has
increased by 12.5%.

Now ulltoa() takes 8% essentially in the reciprocal multiply to divide
by 10. Maybe we can avoid to emit some numbers or only emit them upon
request, since most of the time they're not used.