OPTIM: keep strchr() away from most common request using GET /
strchr() eats ~8% CPU just to look for the '?'. Most of the time there
is none ("GET / HTTP/1.1") or there is one at a well-known place
("GET /? HTTP/1.1"). We can easily recognize these ones by checking
that the request starts with "GET / " and avoid the strchr(), or that
it starts with "GET /?" and avoid it as well, setting the pointer.
This brings almost 5% extra performance on short requests.