MEDIUM: enforce the default maxconn to values retrieved using getrlimit()
In some environments (often containers), the rlim_max value is often as
high as 2^30 (roughly one billion), and we cannot afford to try to allocate
that many FDs, so let's enforce the default maxconn on the value retrieved
this way (roughly 1 million now). This doesn't prevent the user from setting
maxconn in the config or using -n on the command line though.

Thanks to Sasha Nedvedicky for analyzing and reporting this problem.