MINOR: avoid multiplies in calloc() calls
There's always a risk of integer overflow when multiplying the size
by the number of objects like we've done in calloc() calls that are
based on the global.maxsock value. Better pass the maxsock part in
the nmemb argument so that calloc() can check for overflows.

Thanks to Sasha Nedvedicky for reporting this issue.