BUG/MINOR: check and report allocation failures due to too large global.maxsock
We used not to check for particularly inconsistent allocation sizes
because for a long time they used to be deliberate. But now that the
number of FDs is automatically taken from ulimit -Hn, users are seeing
ugly crashes in some environments with too high limits. Let's carefully
add checks for failed allocations that depend on global.maxsock and
provide an informative enough message suggesting to check ulimit -Hn
and/or set a lower value using -n.

Thanks to Sasha Nedvedicky for reporting this issue.