MINOR: support automatic multi-process from bound CPUs on Linux
When -a is added to -D, and -L is in use, we check in /proc/self/status
the list of CPUs the process is allowed to run on, we count them and we
create one less extra process so that as many processes start. No extra
binding is applied, but it is sufficient to start enough workers on
a bunch of CPUs by just using taskset. As a shortcut, -Da also stands
for -D -a, and -DaL stands for -D -a -L.