MEDIUM: report worker num in warnings and alerts
Historically there was something vaguely looking like a date (day of
year, hour, min, sec) that nobody even developers knew how to decode.
Better replace this with the worker number, that's much more useful,
e.g. when dealing with startup issues with -D -a. For example:

  [WARNING] worker 67/80 (288145) : Failed to change pipe size to 81920. Check ulimit -p, fs.pipe-max-size and fs.pipe-user-pages-soft.
  [WARNING] worker 67/80 (288145) : Disabling splice().

suggests that if other workers could start, probably a hard total limit
was reached.